Move `TransformExecutor` class and its methods to `src/step-executors/transform-executor.ts`
* Import `ExpressionEvaluator` and `ReferenceResolver` in `src/step-executors/transform-executor.ts`
* Update import paths for `TransformExecutor` in `src/flow-executor.ts`, `src/step-executors/types.ts`, and `src/types.ts`
